backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx

passlib
bcrypt==4.0.1
python-multipart

python-jose[cryptography] 
python-multipart